JumpState: timeout is set later, not in ctor
Converted setup_timeout(), which was called during initialization and took
a class member as initialization parameter, to set_timeout(), which takes a
time and a destination state as input. The reason will become obvious after
the next commit. Rather than defining the timeout as a property of the state,
we want to model it as what it is, a transition.