MultiSubFSMJumpState: do not wait on fail, error handling
The failed condition now fires already if one sub-FSM failed and it does
not wait for all FSMs to finish (successfully or failing) anymore.
Otherwise we could wait forever, if one FSM failed because the very
reason was that the other FSM did not do something in time, possibly
because it got stuck.

Additionally improved error handling. Messages were upstreamed too
early, when the message had not yet arrived in the expected place.